8343629: More MergeStore benchmark
